+++ crates/web-sys/src/features/mod.rs
@@ -2063,6 +2063,13 @@ mod gen_DomWindowResizeEventDetail;
 #[allow(unused_imports)]
 pub use gen_DomWindowResizeEventDetail::*;
 
+#[cfg(feature = "DoubleRange")]
+#[allow(non_snake_case)]
+mod gen_DoubleRange;
+#[cfg(feature = "DoubleRange")]
+#[allow(unused_imports)]
+pub use gen_DoubleRange::*;
+
 #[cfg(feature = "DragEvent")]
 #[allow(non_snake_case)]
 mod gen_DragEvent;
@@ -4877,6 +4884,13 @@ mod gen_ImageTrackList;
 #[allow(unused_imports)]
 pub use gen_ImageTrackList::*;
 
+#[cfg(feature = "InputDeviceInfo")]
+#[allow(non_snake_case)]
+mod gen_InputDeviceInfo;
+#[cfg(feature = "InputDeviceInfo")]
+#[allow(unused_imports)]
+pub use gen_InputDeviceInfo::*;
+
 #[cfg(feature = "InputEvent")]
 #[allow(non_snake_case)]
 mod gen_InputEvent;
@@ -5583,6 +5597,13 @@ mod gen_MediaStreamTrackState;
 #[cfg(feature = "MediaStreamTrackState")]
 #[allow(unused_imports)]
 pub use gen_MediaStreamTrackState::*;
+
+#[cfg(feature = "MediaTrackCapabilities")]
+#[allow(non_snake_case)]
+mod gen_MediaTrackCapabilities;
+#[cfg(feature = "MediaTrackCapabilities")]
+#[allow(unused_imports)]
+pub use gen_MediaTrackCapabilities::*;
 
 #[cfg(feature = "MediaTrackConstraintSet")]
 #[allow(non_snake_case)]
@@ -9615,6 +9636,13 @@ mod gen_U2fClientData;
 #[cfg(feature = "U2fClientData")]
 #[allow(unused_imports)]
 pub use gen_U2fClientData::*;
+
+#[cfg(feature = "ULongRange")]
+#[allow(non_snake_case)]
+mod gen_ULongRange;
+#[cfg(feature = "ULongRange")]
+#[allow(unused_imports)]
+pub use gen_ULongRange::*;
 
 #[cfg(feature = "UdpMessageEventInit")]
 #[allow(non_snake_case)]