+++ CHANGELOG.md
@@ -17,6 +17,9 @@
 * Added support for returning `Vec`s from async functions.
   [#3630](https://github.com/rustwasm/wasm-bindgen/pull/3630)
 
+* Added bindings for `InputDeviceInfo` and `MediaTrackCapabilities`.
+  [#3935](https://github.com/rustwasm/wasm-bindgen/pull/3935)
+
 * Add bindings for `RTCRtpReceiver.getCapabilities(DOMString)` method.
   [#3941](https://github.com/rustwasm/wasm-bindgen/pull/3941)
 
@@ -38,6 +41,9 @@
 
 * Fix MSRV compilation.
   [#3927](https://github.com/rustwasm/wasm-bindgen/pull/3927)
+
+* Fixed `clippy::empty_docs` lint.
+  [#3946](https://github.com/rustwasm/wasm-bindgen/pull/3946)
 
 --------------------------------------------------------------------------------